feat(iroh): export path and config related tooling
